Update use-dmarc-to-validate-email.md
Please replace "EOP must be the first entry in your domain's MX records in order for DMARC failures to be enforced for your domain." with "EOP does not have to be the first entry for DMARC validation to be done. It just ensures the validation, as we cannot be certain that all on-premise/non-O365 servers will do DMARC checks.  DMARC is eligible to be enforced for a customer’s domain (not server) when you set up the DMARC TXT record, but it is up to the receiving server to actually do the enforcement.  If you set up EOP as the receiving server, then EOP does the DMARC enforcement.", because the first statement is incorrect. Edit approved by SPM Murali Natarajan. Thanks!