delete-items-in-the-recoverable-items-folder-of-mailboxes-on-hold.md
After removing any kind of Hold from the mailbox, the DelayHoldApplied attribute will be set to True, this is a safety net in case Admins accidentally remove Litigation hold from mailboxes. As long as it's set to True the mailbox will still be considered under Hold. 
To remove this run the following cmdlet: 
Set-mailbox -identity user@contoso.com -RemoveDelayHoldApplied